feat: renaming interface for Roles -> Role and adding Team interface
