+++ src/contracts/Team.ts
@@ -1,0 +1,4 @@
+export interface Team {
+  id: number;
+  name: string;
+}
